Add position mapping module
