+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod compression;
 pub mod material_key;
+pub mod position_map;
 pub mod score;
 pub mod table_builder;
 pub mod wdl_file;
